island: Move `get_mode()` and `set_mode()` to a single file `island_mode_service.py`
