--- monkey/monkey_island/cc/services/mode/get_island_mode_service.py
@@ -1,15 +1,0 @@
-from monkey_island.cc.models.island_mode_model import IslandMode
-
-
-def get_mode() -> str:
-    if IslandMode.objects:
-        mode = IslandMode.objects[0].mode
-        return mode
-    else:
-        raise ModeNotSetError
-
-
-class ModeNotSetError(Exception):
-    """
-    Throw this exception when island mode is not set.
-    """
